fix: make `directory` and `multiple` default to 0
Currently, `directory` and `multiple` are left uninitialized in
`method_open_file()`. This leads to them defaulting to `true` (most of
the time – the real behavior is undefined) when the portal client omits
those options in the D-Bus call. This changes them to correctly default
to 0 instead. For reference, the program I ran into this bug with is
VSCodium.